+++ src/containers/index.js
@@ -6,3 +6,5 @@ export { default as ProjectList } from './ProjectList/ProjectList';
 export { default as ProjectEditForm } from './ProjectEditForm/ProjectEditForm';
 export { default as PortfolioProfile } from './PortfolioProfile/PortfolioProfile';
 export { default as PortfolioContents } from './PortfolioContents/PortfolioContents';
+export { default as PortfolioEditProfile } from './PortfolioEditProfile/PortfolioEditProfile';
+export { default as PortfolioEditContents } from './PortfolioEditContents/PortfolioEditContents';
